chore: extract Card sample to separate file for reusability
