+++ packages/react-card/src/stories/CardAppearance.stories.tsx
@@ -1,18 +1,10 @@
 import * as React from 'react';
 
 import { action } from '@storybook/addon-actions';
-import { Body, Caption, Headline } from '@fluentui/react-text';
+import { Headline } from '@fluentui/react-text';
 import { makeStyles, shorthands } from '@griffel/react';
 
-import { Card, CardHeader } from '../index'; // codesandbox-dependency: @fluentui/react-card ^9.0.0-beta
-import { CardProps } from '../components/Card/index';
-import { CardFooter } from '../components/CardFooter/index';
-import { Button } from '@fluentui/react-button';
-import { Open16Regular, Share16Regular } from '@fluentui/react-icons';
-
-const ASSET_URL = 'https://raw.githubusercontent.com/microsoft/fluentui/master/packages/react-card';
-
-const powerpointLogoURL = ASSET_URL + '/assets/powerpoint_logo.svg';
+import SampleCard from './SampleCard';
 
 const useStyles = makeStyles({
   container: {
@@ -23,9 +15,6 @@ const useStyles = makeStyles({
   header: {
     marginBottom: '12px',
   },
-  card: {
-    width: '240px',
-  },
 });
 
 const Title = (props: { children: React.ReactNode }) => {
@@ -35,31 +24,6 @@ const Title = (props: { children: React.ReactNode }) => {
     <Headline as="h4" block className={styles.header}>
       {props.children}
     </Headline>
-  );
-};
-
-const SampleCard = (props: CardProps) => {
-  const styles = useStyles();
-
-  return (
-    <Card className={styles.card} {...props}>
-      <CardHeader
-        image={<img src={powerpointLogoURL} alt="Microsoft PowerPoint logo" />}
-        header={
-          <Body>
-            <b>App Name</b>
-          </Body>
-        }
-        description={<Caption>Developer</Caption>}
-      />
-      Donut chocolate bar oat cake. Dragée tiramisu lollipop bear claw. Marshmallow pastry jujubes toffee sugar plum.
-      <CardFooter>
-        <Button appearance="primary" icon={<Open16Regular />}>
-          Open
-        </Button>
-        <Button icon={<Share16Regular />}>Share</Button>
-      </CardFooter>
-    </Card>
   );
 };
